c: app layout using custom player
